+++ test/console/output/serialized.rb
@@ -41,8 +41,11 @@ describe Console::Output::Serialized do
 			
 			expect(record).to have_keys(
 				subject: be == subject.name,
+				process_id: be == Process.pid,
+				
+				# Backwards compatibility:
 				pid: be == Process.pid,
-				oid: be == subject.object_id,
+				
 				fiber_id: be == Fiber.current.object_id,
 				event: have_keys(
 					type: be == "spawn",